Move PlayerInputManager to own file.
